refactor(lotto): 로또 번호 중복 유효성 검사 로직 추가
- 로또 번호 입력 시 중복된 숫자가 포함될 경우 예외를 발생시키는 유효성 검사 로직을 추가했습니다.
