+++ src/main/java/lotto/domain/Lotto.java
@@ -1,12 +1,18 @@
 package lotto.domain;
 
+import lotto.validation.Validation;
+import lotto.validation.ValidationType;
+
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 public class Lotto {
     private final List<Integer> numbers;
 
     public Lotto(List<Integer> numbers) {
         validate(numbers);
+        Validation.validateLottoDuplicate(numbers);
         this.numbers = numbers;
     }
 
@@ -15,6 +21,7 @@ public class Lotto {
             throw new IllegalArgumentException("[ERROR] 로또 번호는 6개여야 합니다.");
         }
     }
+
     public List<Integer> lottoNumbers() {
         return numbers;
     }
